+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/dcql/DCQLQueryResult.kt
@@ -1,8 +1,0 @@
-package at.asitplus.openid.dcql
-
-data class DCQLQueryResult<Credential: Any>(
-    val credentialQueryMatches: Map<DCQLCredentialQueryIdentifier, List<DCQLCredentialSubmissionOption<Credential>>>
-)
-
-
-