+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/CredentialMatchingResult.kt
@@ -1,20 +1,21 @@
 package at.asitplus.wallet.lib.openid
 
-import at.asitplus.dif.ConstraintField
-import at.asitplus.jsonpath.core.NodeList
-import at.asitplus.openid.dcql.DCQLQueryResult
+import at.asitplus.wallet.lib.agent.HolderDCQLQueryMatchingResult
+import at.asitplus.wallet.lib.agent.HolderPresentationExchangeQueryMatchingResult
+import at.asitplus.wallet.lib.agent.HolderPresentationRequestMatchingResult
 import at.asitplus.wallet.lib.data.CredentialPresentationRequest
 
 sealed interface CredentialMatchingResult<Credential: Any> {
     val presentationRequest: CredentialPresentationRequest
+    val matchingResult: HolderPresentationRequestMatchingResult<Credential>
 }
 
 data class PresentationExchangeMatchingResult<Credential: Any>(
     override val presentationRequest: CredentialPresentationRequest.PresentationExchangeRequest,
-    val matchingInputDescriptorCredentials: Map<String, Map<Credential, Map<ConstraintField, NodeList>>>,
+    override val matchingResult: HolderPresentationExchangeQueryMatchingResult<Credential>,
 ) : CredentialMatchingResult<Credential>
 
 data class DCQLMatchingResult<Credential: Any>(
     override val presentationRequest: CredentialPresentationRequest.DCQLRequest,
-    val dcqlQueryResult: DCQLQueryResult<Credential>,
+    override val matchingResult: HolderDCQLQueryMatchingResult<Credential>,
 ) : CredentialMatchingResult<Credential>