+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/HolderPresentationRequestMatchingResult.kt
@@ -1,0 +1,5 @@
+package at.asitplus.wallet.lib.agent
+
+interface HolderPresentationRequestMatchingResult<Credential: Any> {
+    val credentials: List<Credential>
+}